Use float in font_descr

+++ font_descr.hpp
@@ -7,19 +7,19 @@ namespace rematrix {
 
 struct glyph_descr
 {
-    const array<double, 2> pos;
-    const array<double, 2> size;
+    const array<float, 2> pos;
+    const array<float, 2> size;
 };
 
 struct font_descr
 {
-    array<array<double, 2>, 4>
+    array<array<float, 2>, 4>
     tex_coords(char c) const
     {
         return tex_coords(glyphs.at(c));
     }
 
-    array<array<double, 2>, 4>
+    array<array<float, 2>, 4>
     tex_coords(const glyph_descr& g) const
     {
         auto [x, y] = g.pos;
